requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
chardet==5.2.0
Django==4.2
environs==11.2.1
frozenlist==1.5.0
idna==3.10
imageio==2.36.1
magic-filter==1.0.12
marshmallow==3.23.1
multidict==6.1.0
numpy==2.0.2
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.25.1
python-dotenv==1.0.1
reportlab==4.2.5
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.18.3
